+++ myems-web/src/components/MyEMS/common/ChildSpacesTable.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import BootstrapTable from 'react-bootstrap-table-next';
 import FalconCardHeader from '../../common/FalconCardHeader';
-import { Card, CardBody, CardHeader, Row, Col } from 'reactstrap';
+import { Card, CardBody, Row, Col } from 'reactstrap';
 import { withTranslation } from 'react-i18next';